进一步优化节点信息，使用原子操作修改free属性

+++ nsqlookupd/http.go
@@ -267,7 +267,7 @@ type node struct {
 	Version          string   `json:"version"`
 	Tombstones       []bool   `json:"tombstones"`
 	Topics           []string `json:"topics"`
-	Free             int
+	Free             int64
 	IpAddress        string
 }
